Create SyntheticControlWeightedMeanDifferences
The goal here is to test the validity of a synthetic control and demonstrate how the control is similar to the treated on observable characteristics. 

First, use the Synth package in R to obtain the synthetic control weights. Then, extract the weights and compare unweighted and weighted differences across important characteristics