Create Synthetic-Control
We had a population of treated companies, with tens of thousands of potential companies in the donor pool.  To avoid overfitting, I took random draws of the data and then formed SCs from the top orgs remaining in the draw that were "closest" to the treated group past average on the outcome (probably could have done propensity scores as well). I then plot the distribution of the SC results.

One question that came up: Is it better to show this distribution of SCs, or instead to use a training set to get the "optimal" SC, and then use that going forward?  This can be tested!